fix tests for Profiles class
